refactor: moved model responses file to azul folders (#714)
