+++ explorer/app/apis/azul/common/entities.ts
@@ -24,16 +24,3 @@ export type AzulEntitiesResponses =
   | DonorsResponse
   | FileEntityResponse
   | LibrariesResponse;
-
-/**
- * Model of file "leaf" values in matrix tree response from Azul.
- */
-export interface ProjectMatrixFileResponse {
-  contentDescription: string[];
-  matrixCellCount?: number;
-  name: string;
-  size: number;
-  url: string;
-  uuid: string;
-  version: string;
-}
