+++ explorer/app/apis/azul/hca-dcp/common/entities.ts
@@ -1,3 +1,31 @@
+/**
+ * Model of accessions included in the response from index/projects or /projects/uuid API endpoints.
+ */
+export interface AccessionResponse {
+  accession: string;
+  namespace: string;
+}
+
+/**
+ * Model of contributor value included in the response from index/projects API endpoint.
+ */
+export interface ContributorResponse {
+  contactName: string;
+  correspondingContributor?: boolean;
+  email?: string;
+  institution: string;
+  laboratory?: string;
+  projectRole?: string;
+}
+
+/**
+ * Model of fileFormat value included in the response from index/summary API endpoint.
+ */
+export interface FileFormatResponse {
+  count: number;
+  format: string;
+}
+
 export interface FilesEntity {
   contentDescription: string[];
   format: string;
@@ -9,6 +37,84 @@ export interface FilesEntity {
 
 export interface FilesEntityResponse {
   files: FilesEntity[];
+}
+
+/**
+ * Model of file "leaf" values in matrix tree response from Azul.
+ */
+export interface ProjectMatrixFileResponse {
+  contentDescription: string[];
+  matrixCellCount?: number;
+  name: string;
+  size: number;
+  url: string;
+  uuid: string;
+  version: string;
+}
+
+/**
+ * Model of project value nested in response returned from index/projects API endpoint.
+ */
+export interface ProjectResponse {
+  accessions: AccessionResponse[];
+  contributedAnalyses: ProjectResponseContributedAnalyses;
+  contributors: ContributorResponse[];
+  estimatedCellCount: number;
+  matrices: ProjectResponseMatrices;
+  projectDescription: string;
+  projectId: string;
+  projectShortname: string;
+  projectTitle: string;
+  publications: PublicationResponse[];
+  supplementaryLinks: string[];
+}
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any -- TODO - revisit nested Azul structure.
+export type ProjectResponseContributedAnalyses = { [key: string]: unknown };
+// eslint-disable-next-line @typescript-eslint/no-explicit-any -- TODO - revisit nested Azul structure.
+export type ProjectResponseMatrices = { [key: string]: unknown };
+
+/**
+ * Model of response returned from /index/projects API endpoint.
+ */
+export interface ProjectsResponse {
+  cellSuspensions?: {
+    totalCells: number;
+  }[];
+  donorOrganisms: {
+    developmentStage: string[];
+    disease: string[];
+    donorCount: number;
+    genusSpecies: string[];
+  }[];
+  entryId: string;
+  fileTypeSummaries: {
+    count: number;
+    format: string;
+  }[];
+  projects: ProjectResponse[];
+  protocols: {
+    libraryConstructionApproach?: string[];
+    nucleicAcidSource?: string[];
+    pairedEnd?: boolean[];
+    workflow?: string[];
+  }[];
+  samples: {
+    disease: string[];
+    organ: string[];
+    organPart: string[];
+    sampleEntityType: string[];
+  }[];
+}
+
+/**
+ * Model of publication value nested in project response returned from index/projects API endpoint.
+ */
+export interface PublicationResponse {
+  doi: string | null;
+  officialHcaPublication: boolean | null;
+  publicationTitle: string;
+  publicationUrl: string;
 }
 
 export interface SamplesEntity {
