+++ explorer/app/hooks/useProjectMatrixArchiveFiles.ts
@@ -4,7 +4,7 @@ import { useEffect } from "react";
 import {
   ArchiveFile,
   fetchProjectMatrixArchiveFiles,
-} from "../apis/azul/hca-dcp/common/archivePreview";
+} from "../apis/azul/hca-dcp/archivePreview/archivePreview";
 import { ProjectMatrixView } from "../viewModelBuilders/azul/hca-dcp/common/projectMatrixMapper/entities";
 
 export interface UseArchivePreview {
