+++ explorer/app/transformers/hca.ts
@@ -1,8 +1,9 @@
+import { ProjectsResponse } from "../apis/azul/hca-dcp/common/entities";
+
 /**
  * Entity transformers. Used to have a configurable way to transform different response models into a
  * single view model expected by container components.
  */
-import { ProjectsResponse } from "../models/responses";
 
 /**
  * Determine the ID of the given detail type.
