+++ explorer/app/viewModelBuilders/azul/hca-dcp/common/projectMatrixMapper/projectMatrixMapper.ts
@@ -1,10 +1,10 @@
 import { HCA_DCP_CATEGORY_KEY } from "../../../../../../site-config/hca-dcp/category";
-import { ProjectMatrixFileResponse } from "../../../../../apis/azul/common/entities";
 import { isNullOrUndefined } from "../../../../../apis/azul/common/utils";
 import {
+  ProjectMatrixFileResponse,
   ProjectResponseContributedAnalyses,
   ProjectResponseMatrices,
-} from "../../../../../models/responses";
+} from "../../../../../apis/azul/hca-dcp/common/entities";
 import { ProjectMatrixTableView, ProjectMatrixView } from "./entities";
 
 // Template constants
