+++ explorer/site-config/hca-dcp/dev/detail/project/top.ts
@@ -2,9 +2,9 @@ import {
   ComponentConfig,
   ComponentsConfig,
 } from "@clevercanary/data-explorer-ui/lib/config/entities";
-import { ProjectsResponse } from "app/models/responses";
+import { ProjectsResponse } from "../../../../../app/apis/azul/hca-dcp/common/entities";
 import * as C from "../../../../../app/components";
-import * as T from "../../projectViewModelBuilder";
+import * as T from "../../../../../app/viewModelBuilders/azul/hca-dcp/common/projectViewModelBuilder";
 
 export const top: ComponentsConfig = [
   {
