+++ explorer/site-config/hca-dcp/dev/index/projectsEntityConfig.ts
@@ -4,18 +4,10 @@ import {
   ListConfig,
   SORT_DIRECTION,
 } from "@clevercanary/data-explorer-ui/lib/config/entities";
-import { ProjectsResponse } from "app/models/responses";
 import { getProjectId } from "app/transformers/hca";
-import * as Components from "../../../../app/components";
-import {
-  projectsBuildAnatomicalEntityColumn,
-  projectsBuildCellCountColumn,
-  projectsBuildDevelopmentStage,
-  projectsBuildDiseaseDonorColumn,
-  projectsBuildLibraryConstructionApproachColumn,
-  projectsBuildProjectTitleColumn,
-  projectsBuildSpecies,
-} from "../../../../app/viewModelBuilders/azul/hca-dcp/common/viewModelBuilders";
+import { ProjectsResponse } from "../../../../app/apis/azul/hca-dcp/common/entities";
+import * as C from "../../../../app/components";
+import * as V from "../../../../app/viewModelBuilders/azul/hca-dcp/common/viewModelBuilders";
 import { HCA_DCP_CATEGORY_KEY, HCA_DCP_CATEGORY_LABEL } from "../../category";
 import { PROJECTS_LABEL } from "../constants";
 import { mainColumn as exportMainColumn } from "../detail/project/exportMainColumn";
@@ -76,17 +68,17 @@ export const projectsEntityConfig: EntityConfig = {
     columns: [
       {
         componentConfig: {
-          component: Components.Link,
-          viewBuilder: projectsBuildProjectTitleColumn,
-        } as ComponentConfig<typeof Components.Link>,
+          component: C.Link,
+          viewBuilder: V.projectsBuildProjectTitleColumn,
+        } as ComponentConfig<typeof C.Link>,
         header: HCA_DCP_CATEGORY_LABEL.PROJECT_TITLE,
         id: HCA_DCP_CATEGORY_KEY.PROJECT_TITLE,
         width: { max: "2fr", min: "374px" },
       },
       {
         componentConfig: {
-          component: Components.NTagCell,
-          viewBuilder: projectsBuildSpecies,
+          component: C.NTagCell,
+          viewBuilder: V.projectsBuildSpecies,
         },
         header: "Species", // TODO confirm header
         id: HCA_DCP_CATEGORY_KEY.GENUS_SPECIES,
@@ -94,8 +86,8 @@ export const projectsEntityConfig: EntityConfig = {
       },
       {
         componentConfig: {
-          component: Components.NTagCell,
-          viewBuilder: projectsBuildLibraryConstructionApproachColumn,
+          component: C.NTagCell,
+          viewBuilder: V.projectsBuildLibraryConstructionApproachColumn,
         },
         header: HCA_DCP_CATEGORY_LABEL.LIBRARY_CONSTRUCTION_METHOD,
         id: HCA_DCP_CATEGORY_KEY.LIBRARY_CONSTRUCTION_METHOD,
@@ -103,8 +95,8 @@ export const projectsEntityConfig: EntityConfig = {
       },
       {
         componentConfig: {
-          component: Components.NTagCell,
-          viewBuilder: projectsBuildAnatomicalEntityColumn,
+          component: C.NTagCell,
+          viewBuilder: V.projectsBuildAnatomicalEntityColumn,
         },
         header: HCA_DCP_CATEGORY_LABEL.ANATOMICAL_ENTITY,
         id: HCA_DCP_CATEGORY_KEY.ANATOMICAL_ENTITY,
@@ -112,8 +104,8 @@ export const projectsEntityConfig: EntityConfig = {
       },
       {
         componentConfig: {
-          component: Components.NTagCell,
-          viewBuilder: projectsBuildDiseaseDonorColumn,
+          component: C.NTagCell,
+          viewBuilder: V.projectsBuildDiseaseDonorColumn,
         },
         header: "Disease (Donor)", // TODO confirm header
         id: HCA_DCP_CATEGORY_KEY.DONOR_DISEASE,
@@ -123,15 +115,15 @@ export const projectsEntityConfig: EntityConfig = {
         componentConfig: {
           children: [
             {
-              component: Components.Cell,
-              viewBuilder: projectsBuildCellCountColumn,
-            } as ComponentConfig<typeof Components.Cell>,
+              component: C.Cell,
+              viewBuilder: V.projectsBuildCellCountColumn,
+            } as ComponentConfig<typeof C.Cell>,
           ],
-          component: Components.Tooltip,
+          component: C.Tooltip,
           props: {
             title: "Cell Count Estimate",
           },
-        } as ComponentConfig<typeof Components.Tooltip>,
+        } as ComponentConfig<typeof C.Tooltip>,
         header: HCA_DCP_CATEGORY_LABEL.EFFECTIVE_CELL_COUNT,
         id: HCA_DCP_CATEGORY_KEY.EFFECTIVE_CELL_COUNT,
         width: { max: "1fr", min: "96px" },
@@ -139,9 +131,9 @@ export const projectsEntityConfig: EntityConfig = {
       {
         columnVisible: false,
         componentConfig: {
-          component: Components.NTagCell,
-          viewBuilder: projectsBuildDevelopmentStage,
-        } as ComponentConfig<typeof Components.NTagCell>,
+          component: C.NTagCell,
+          viewBuilder: V.projectsBuildDevelopmentStage,
+        } as ComponentConfig<typeof C.NTagCell>,
         header: HCA_DCP_CATEGORY_LABEL.DEVELOPMENT_STAGE,
         id: HCA_DCP_CATEGORY_KEY.DEVELOPMENT_STAGE,
         width: { max: "1fr", min: "148px" },
