+++ explorer/site-config/lungmap/dev/detail/project/matricesMainColumn.ts
@@ -1,7 +1,7 @@
 import { ComponentConfig } from "@clevercanary/data-explorer-ui/lib/config/entities";
+import { ProjectsResponse } from "../../../../../app/apis/azul/hca-dcp/common/entities";
 import * as C from "../../../../../app/components";
 import * as MDX from "../../../../../app/content/lungmap";
-import { ProjectsResponse } from "../../../../../app/models/responses";
 import * as T from "../../../../../app/viewModelBuilders/azul/lungmap/common/viewModelBuilders";
 
 export const mainColumn: ComponentConfig[] = [
